tests/test_auth.py
"""
Test suite for authentication module.
"""
import pytest
import jwt
from datetime import datetime, timedelta, timezone
from unittest.mock import patch, MagicMock
from flask import request, Flask
from werkzeug.security import generate_password_hash

from app.models import User, db
from app.auth import generate_token, verify_token, token_required, admin_required, init_app

@pytest.fixture(scope='function')
def app():
    """Create and configure a test Flask application."""
    app = Flask(__name__)
    app.config['SECRET_KEY'] = 'test-secret-key'
    app.config['TESTING'] = True
    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False

    with app.app_context():
        db.init_app(app)
        db.create_all()
        yield app
        db.session.remove()
        db.drop_all()

@pytest.fixture
def client(app):
    """Create a test client."""
    return app.test_client()

@pytest.fixture
def test_user(app):
    """Create a test user."""
    with app.app_context():
        user = User(
            username='testuser',
            password_hash=generate_password_hash('password123'),
            role='user'
        )
        db.session.add(user)
        db.session.commit()

        # Get a fresh instance to avoid detached session issues
        user = User.query.filter_by(username='testuser').first()
        yield user

        db.session.rollback()

@pytest.fixture
def test_admin(app):
    """Create a test admin user."""
    with app.app_context():
        admin = User(
            username='admin',
            password_hash=generate_password_hash('admin123'),
            role='admin'
        )
        db.session.add(admin)
        db.session.commit()

        # Get a fresh instance to avoid detached session issues
        admin = User.query.filter_by(username='admin').first()
        yield admin

        db.session.rollback()

def test_generate_token_success(app, test_user):
    """Test successful token generation."""
    with app.app_context():
        # Re-query the user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)
        assert isinstance(token, str)

        payload = jwt.decode(
            token,
            app.config['SECRET_KEY'],
            algorithms=['HS256']
        )
        assert payload['username'] == user.username
        assert payload['role'] == user.role
        assert all(key in payload for key in ['exp', 'iat', 'nbf'])

def test_generate_token_user_not_found(app):
    """Test token generation with non-existent user."""
    with app.app_context():
        with pytest.raises(ValueError, match='User not found'):
            generate_token('nonexistent_user')

def test_generate_token_database_error(app):
    """Test token generation with database error."""
    with app.app_context():
        with patch('app.models.User.query') as mock_query:
            mock_query.filter_by.side_effect = Exception('Database error')
            with pytest.raises(Exception):
                generate_token('testuser')

def test_verify_token_success(app, test_user):
    """Test successful token verification."""
    with app.app_context():
        # Re-query the user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)
        payload = verify_token(token)
        assert payload is not None
        assert payload['username'] == user.username
        assert payload['role'] == user.role

def test_token_required_valid(app, client, test_user):
    """Test token_required decorator with valid token."""
    with app.app_context():
        # Re-query the user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)

        @app.route('/protected')
        @token_required
        def protected_route():
            return {'message': 'success', 'user': request.current_user.username}

        response = client.get('/protected', headers={'Authorization': f'Bearer {token}'})
        assert response.status_code == 200
        assert response.json['user'] == user.username

def test_token_required_missing_header(app, client):
    """Test token_required without Authorization header."""
    with app.app_context():
        @app.route('/protected-no-auth')
        @token_required
        def protected_route():
            return {'message': 'success'}

        response = client.get('/protected-no-auth')
        assert response.status_code == 401
        assert response.json['message'] == 'Token is missing'

def test_token_required_invalid_format(app, client):
    """Test token_required with invalid token format."""
    with app.app_context():
        @app.route('/protected-invalid')
        @token_required
        def protected_route():
            return {'message': 'success'}

        headers = {'Authorization': 'InvalidFormat'}
        response = client.get('/protected-invalid', headers=headers)
        assert response.status_code == 401
        assert response.json['message'] == 'Invalid token format'

def test_admin_required_success(app, client, test_admin):
    """Test admin_required decorator with admin user."""
    with app.app_context():
        # Re-query the admin to ensure it's attached to the session
        admin = User.query.filter_by(username=test_admin.username).first()
        token = generate_token(admin.username)

        @app.route('/admin-only')
        @admin_required
        def admin_route():
            return {'message': 'success'}

        response = client.get('/admin-only', headers={'Authorization': f'Bearer {token}'})
        assert response.status_code == 200

def test_admin_required_non_admin(app, client, test_user):
    """Test admin_required decorator with non-admin user."""
    with app.app_context():
        # Re-query the user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)

        @app.route('/admin-only-2')
        @admin_required
        def admin_route():
            return {'message': 'success'}

        response = client.get('/admin-only-2', headers={'Authorization': f'Bearer {token}'})
        assert response.status_code == 403
        assert response.json['message'] == 'Admin privileges required'

def test_token_required_unexpected_error(app, client, test_user):
    """Test token_required decorator with unexpected error during verification."""
    with app.app_context():
        token = generate_token(test_user.username)

        @app.route('/protected-error')
        @token_required
        def protected_route():
            return {'message': 'success'}

        with patch('app.auth.verify_token', side_effect=Exception('Unexpected error')):
            response = client.get('/protected-error',
                                headers={'Authorization': f'Bearer {token}'})
            assert response.status_code == 401
            assert 'Token validation failed: Unexpected error' == response.json['message']

def test_init_app_with_secret_key():
    """Test init_app with existing secret key."""
    app = Flask(__name__)
    app.config['SECRET_KEY'] = 'existing-secret'
    init_app(app)
    assert app.config['SECRET_KEY'] == 'existing-secret'

def test_init_app_without_secret_key():
    """Test init_app without existing secret key."""
    app = Flask(__name__)
    init_app(app)
    assert app.config['SECRET_KEY'] == 'your-secret-key-here'

def test_generate_token_secret_key_error(app, test_user):
    """Test token generation with secret key error."""
    with app.app_context():
        with patch('flask.current_app.config.get') as mock_config:
            mock_config.side_effect = Exception('Config error')
            with pytest.raises(Exception) as exc_info:
                generate_token(test_user.username)
            assert 'Config error' in str(exc_info.value)

def test_generate_token_jwt_encode_error(app, test_user):
    """Test token generation with JWT encode error."""
    with app.app_context():
        with patch('jwt.encode') as mock_encode:
            mock_encode.side_effect = Exception('Encoding error')
            with pytest.raises(Exception) as exc_info:
                generate_token(test_user.username)
            assert 'Encoding error' in str(exc_info.value)

def test_token_required_split_error(app, client):
    """Test token_required decorator with malformed authorization header."""
    with app.app_context():
        @app.route('/protected-split-error')
        @token_required
        def protected_route():
            return {'message': 'success'}

        response = client.get('/protected-split-error',
                            headers={'Authorization': 'Bearer'})
        assert response.status_code == 401
        # Changed to match actual implementation
        assert response.json['message'] == 'Invalid token format'

def test_token_required_header_error(app, client):
    """Test token_required decorator with header processing error."""
    with app.app_context():
        @app.route('/protected-header-error')
        @token_required
        def protected_route():
            return {'message': 'success'}

        # Send request without headers
        response = client.get('/protected-header-error')
        assert response.status_code == 401
        assert response.json['message'] == 'Token is missing'

# Update existing tests that had assertion mismatches
def test_verify_token_invalid_signature(app, test_user):
    """Test verification of token with invalid signature."""
    with app.app_context():
        # Re-query user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)
        invalid_token = token[:-5] + 'wrong'  # Modify the signature part
        assert verify_token(invalid_token) is None

def test_verify_token_expired(app, test_user):
    """Test verification of expired token."""
    with app.app_context():
        # Re-query user to ensure it's attached to the session
        user = User.query.filter_by(username=test_user.username).first()
        now = datetime.now(timezone.utc)
        payload = {
            'username': user.username,
            'role': user.role,
            'exp': int((now - timedelta(hours=1)).timestamp()),
            'iat': int(now.timestamp()),
            'nbf': int(now.timestamp())
        }
        token = jwt.encode(payload, app.config['SECRET_KEY'], algorithm='HS256')
        assert verify_token(token) is None

def test_token_required_user_not_found(app, client):
    """Test token_required when user is deleted after token generation."""
    with app.app_context():
        # Create temporary user
        user = User(
            username='temp_user',
            password_hash=generate_password_hash('password123'),
            role='user'
        )
        db.session.add(user)
        db.session.commit()

        # Generate token while user exists
        token = generate_token(user.username)

        # Delete the user
        db.session.delete(user)
        db.session.commit()

        @app.route('/protected-deleted-user')
        @token_required
        def protected_route():
            return {'message': 'success'}

        response = client.get('/protected-deleted-user',
                            headers={'Authorization': f'Bearer {token}'})
        assert response.status_code == 401
        # Changed to match actual implementation
        assert response.json['message'] == 'Invalid token'

def test_verify_token_missing_secret_key(app, test_user):
    """Test token verification with missing secret key."""
    with app.app_context():
        token = generate_token(test_user.username)
        with patch('flask.current_app.config.get', return_value=None):
            assert verify_token(token) is None

def test_verify_token_database_error(app, test_user):
    """Test verify_token with database error during user lookup."""
    with app.app_context():
        token = generate_token(test_user.username)
        # Patch the user query after token payload is decoded
        with patch('app.models.User.query') as mock_query:
            mock_query.filter_by.return_value.first.side_effect = Exception('Database error')
            result = verify_token(token)
            assert result is None

def test_token_required_verify_token_error(app, client):
    """Test token_required when verify_token raises an unexpected error."""
    with app.app_context():
        @app.route('/protected-verify-error')
        @token_required
        def protected_route():
            return {'message': 'success'}

        # Use an invalid token that will cause a verification error
        headers = {'Authorization': 'Bearer invalid.token.here'}
        response = client.get('/protected-verify-error', headers=headers)
        assert response.status_code == 401
        assert response.json['message'] == 'Invalid token'


def test_token_required_user_lookup_error(app, client, test_user):
    """Test token_required when user lookup raises an error."""
    with app.app_context():
        # First, create a valid token
        user = User.query.filter_by(username=test_user.username).first()
        token = generate_token(user.username)

        @app.route('/protected-lookup-error')
        @token_required
        def protected_route():
            return {'message': 'success'}

        # Now patch verify_token to return a valid payload but force user lookup to fail
        with patch('app.auth.verify_token') as mock_verify:
            mock_verify.return_value = {
                'username': user.username,
                'role': user.role
            }
            # We need to patch the User query in the token_required decorator
            with patch('app.models.User.query') as mock_query:
                # Create a mock filter_by that raises an exception
                mock_filter_by = MagicMock()
                mock_filter_by.first.side_effect = Exception('Database error')
                mock_query.filter_by.return_value = mock_filter_by

                response = client.get('/protected-lookup-error',
                                      headers={'Authorization': f'Bearer {token}'})

                assert response.status_code == 401
                assert response.json['message'] == 'Token validation failed: Database error'

                # Verify our mock was called with the correct username
                mock_query.filter_by.assert_called_with(username=user.username)

# Test to cover the error path in verify_token
def test_verify_token_with_error(app, test_user):
    """Test verify_token when an unexpected error occurs."""
    with app.app_context():
        token = generate_token(test_user.username)
        with patch('jwt.decode') as mock_decode:
            mock_decode.side_effect = Exception('Unexpected verification error')
            result = verify_token(token)
            assert result is None

# Test for token_required with verify_token returning None
def test_token_required_verify_returns_none(app, client):
    """Test token_required when verify_token returns None."""
    with app.app_context():
        @app.route('/protected-verify-none')
        @token_required
        def protected_route():
            return {'message': 'success'}

        with patch('app.auth.verify_token', return_value=None):
            response = client.get('/protected-verify-none',
                                headers={'Authorization': 'Bearer some.token.here'})
            assert response.status_code == 401
            assert response.json['message'] == 'Invalid token'